refactored to include camelCasing, backend namespace, refactor
