+++ tests/agraph_backend_tests.cpp
@@ -2,13 +2,15 @@
 #include <vector>
 
 #include <Eigen/Dense>
-#include "gtest/gtest.h"
+#include <gtest/gtest.h>
 
 #include "BingoCpp/backend.h"
 #include "testing_utils.h"
 #include "test_fixtures.h"
 
 using namespace bingo;
+using namespace backend;
+
 namespace {
 const int N_OPS = 13;
 
@@ -53,9 +55,9 @@ TEST_P(AGraphBackend, simplify_and_evaluate) {
   stack << 0, 0, 0,
            0, 1, 0,
            operator_i, 0, 0;
-  Eigen::ArrayXXd f_of_x = simplify_and_evaluate(stack,
-                                                 sample_agraph_1_values.x_vals,
-                                                 sample_agraph_1_values.constants);
+  Eigen::ArrayXXd f_of_x = simplifyAndEvaluate(stack,
+                                               sample_agraph_1_values.x_vals,
+                                               sample_agraph_1_values.constants);
   ASSERT_TRUE(testutils::almost_equal(expected_outcome, f_of_x));
 }
 
@@ -74,10 +76,10 @@ TEST_P(AGraphBackend, simplify_and_evaluate_x_deriv) {
   Eigen::ArrayXXd x_0 = sample_agraph_1_values.x_vals;
   Eigen::ArrayXXd constants = sample_agraph_1_values.constants;
   std::pair<Eigen::ArrayXXd, Eigen::ArrayXXd> res_and_gradient = 
-    simplify_and_evaluate_with_derivative(stack,
-                                          x_0,
-                                          constants,
-                                          true);
+    simplifyAndEvaluateWithDerivative(stack,
+                                      x_0,
+                                      constants,
+                                      true);
   Eigen::ArrayXXd df_dx = res_and_gradient.second;
   ASSERT_TRUE(testutils::almost_equal(expected_derivative, df_dx));
 }
@@ -100,10 +102,10 @@ TEST_P(AGraphBackend, simplify_and_evaluate_c_deriv) {
   Eigen::ArrayXXd x_0 = sample_agraph_1_values.x_vals;
   Eigen::ArrayXXd constants = sample_agraph_1_values.constants;
   std::pair<Eigen::ArrayXXd, Eigen::ArrayXXd> res_and_gradient = 
-    simplify_and_evaluate_with_derivative(stack,
-                                          x_0,
-                                          constants,
-                                          false);
+    simplifyAndEvaluateWithDerivative(stack,
+                                      x_0,
+                                      constants,
+                                      false);
   Eigen::ArrayXXd df_dc = res_and_gradient.second;
   ASSERT_TRUE(testutils::almost_equal(expected_derivative, df_dc));
 }
@@ -119,7 +121,7 @@ TEST_F(AGraphBackend, evaluate) {
 
 TEST_F(AGraphBackend, evaluate_and_derivative) {
   std::pair<Eigen::ArrayXXd, Eigen::ArrayXXd> y_and_dy =
-    evaluate_with_derivative(simple_stack, x, constants);
+    evaluateWithDerivative(simple_stack, x, constants);
   Eigen::ArrayXXd y_true = x.col(0) * (constants[0] + constants[1] 
                           / x.col(1)) - x.col(0);
   Eigen::ArrayXXd dy_true = Eigen::ArrayXXd::Zero(3, 3);
@@ -132,40 +134,21 @@ TEST_F(AGraphBackend, evaluate_and_derivative) {
 
 TEST_F(AGraphBackend, mask_evaluate) {
   Eigen::ArrayXXd y = evaluate(simple_stack, x, constants);
-  Eigen::ArrayXXd y_simple = simplify_and_evaluate(simple_stack, x, constants);
+  Eigen::ArrayXXd y_simple = simplifyAndEvaluate(simple_stack, x, constants);
   ASSERT_TRUE(testutils::almost_equal(y, y_simple));
 }
 
 TEST_F(AGraphBackend, mask_evaluate_and_derivative) {
   std::pair<Eigen::ArrayXXd, Eigen::ArrayXXd> y_and_dy =
-    evaluate_with_derivative(simple_stack, x, constants);
+    evaluateWithDerivative(simple_stack, x, constants);
   std::pair<Eigen::ArrayXXd, Eigen::ArrayXXd> y_and_dy_simple =
-    simplify_and_evaluate_with_derivative(simple_stack, x, constants);
+    simplifyAndEvaluateWithDerivative(simple_stack, x, constants);
   ASSERT_TRUE(testutils::almost_equal(y_and_dy.first, y_and_dy_simple.first));
   ASSERT_TRUE(testutils::almost_equal(y_and_dy.first, y_and_dy_simple.first));
 }
 
-// TEST_F(AcyclicGraphTest, simplify) {
-//   // shorter stack
-//   std::cout << "stack\n" << stack << std::endl;
-//   Eigen::ArrayX3i short_stack = SimplifyStack(stack);
-//   std::cout << "2\n";
-//   ASSERT_LE(short_stack.rows(), stack.rows());
-//   std::cout << "3\n";
-
-//   // equivalent evatuation
-//   Eigen::ArrayXXd y = Evaluate(stack, x, constants);
-//   std::cout << "4\n";
-//   Eigen::ArrayXXd simplified_y = Evaluate(short_stack, x, constants);
-//   std::cout << "5\n";
-
-//   for (size_t i = 0; i < x.rows(); ++i) {
-//     ASSERT_EQ(y(i), simplified_y(i));
-//   }
-// }
-
 TEST_F(AGraphBackend, get_utilized_commands) {
-  std::vector<bool> used_commands = get_utilized_commands(simple_stack);
+  std::vector<bool> used_commands = getUtilizedCommands(simple_stack);
   int num_used_commands = 0;
   for (auto const& command_is_used : used_commands) {
     if (command_is_used) {
